examples/fastAPI/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "haiway_fastapi"
description = "Example of haiway usage with fastapi."
version = "0.1.0"
readme = "README.md"
maintainers = [
  { name = "Kacper Kaliński", email = "kacper.kalinski@miquido.com" },
]
requires-python = ">=3.12"
dependencies = [
  "haiway~=0.1.0",
  "fastapi-slim~=0.115",
  "asyncpg~=0.29",
  "httpx~=0.25.0",
]

[project.urls]
Homepage = "https://miquido.com"

[project.optional-dependencies]
dev = [
  "uvicorn~=0.30",
  "ruff~=0.5.0",
  "pyright~=1.1",
  "bandit~=1.7",
]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [".venv", ".git", ".cache"]
lint.select = ["E", "F", "A", "I", "B", "PL", "W", "C", "RUF", "UP"]
lint.ignore = []

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "E402"]
"./tests/*.py" = ["PLR2004"]

[tool.pyright]
pythonVersion = "3.12"
venvPath = "."
venv = ".venv"
include = ["./src"]
exclude = ["**/node_modules", "**/__pycache__"]
ignore = []
stubPath = "./stubs"
reportMissingImports = true
reportMissingTypeStubs = true
typeCheckingMode = "strict"
userFileIndexingLimit = -1
useLibraryCodeForTypes = true

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
